Add the initial code
This commit updates the initial code for this application. This has been
written as a bit of an excercise to see if FastAPI is a better neater
option for implemeting the back end. Long story it is, so its time to
push up this version and start using git to manage this project.