Add a new `run-tests-native` command to `Makefile`
Prior to this change, there was only an option for running this
application as a docker container. This is the intended way to run the
application however it meant there was a dev time dependancy on having
docker installed and enough spare ram to spin up a virtual machine when
developing on Macos and Windows.

This change adds the `run-tests-native` command which uses the locally
managed rye virtual environment to spin up the application and run the
tests natively. This means that there is no longer a dependency for
docker when developing locally.