Update test suite layout to split unit/integration tests
Prior to this change, there was a test file (`auth/test_usecases.py`)
that touched the database living in the unit tests folder.

This change moves the test file that interacts with the database to the
integration test folder so that we have a clear delineation between
which tests can be ran quickly in pure python land, vs tests that need
database access (and are generally slower/more expensive to run).